refactored common SparkplugB code to uns_mqtt
