+++ .vscode/settings.json
@@ -8,6 +8,7 @@
         "EMQX",
         "graphdb",
         "HIVEMQ",
+        "hostpath",
         "hypertable",
         "IIOT",
         "integrationtest",
@@ -22,6 +23,7 @@
         "NCMD",
         "NDATA",
         "NDEATH",
+        "openebs",
         "packettypes",
         "paho",
         "ppcle",
@@ -34,6 +36,7 @@
         "SCADA",
         "sparkplugb",
         "sslmode",
+        "storageclass",
         "Tahu",
         "timescaledb",
         "unifiednamespace",
