+++ 02_mqtt-cluster/protobuf/include/google/protobuf/empty.proto
