+++ 02_mqtt-cluster/protobuf/include/google/protobuf/struct.proto
