+++ 02_mqtt-cluster/requirements.txt
@@ -1,3 +1,4 @@
 wheel>=0.37.1
-paho-mqtt==1.6.1
 logger>=1.4
+paho-mqtt==1.6.1
+protobuf>=4.21.1
