+++ 02_mqtt-cluster/sparkplug_b/sparkplug_b.proto
