+++ 05_sparkplugb/test/test_sparkplugb_enc_config.py
@@ -12,10 +12,18 @@ cmd_subfolder = os.path.realpath(
         os.path.join(
             os.path.split(inspect.getfile(inspect.currentframe()))[0], '..',
             'src')))
+uns_mqtt_folder = os.path.realpath(
+    os.path.abspath(
+        os.path.join(cmd_subfolder, '..', '..', '02_mqtt-cluster', 'src')))
+
 if cmd_subfolder not in sys.path:
     sys.path.insert(0, cmd_subfolder)
-    sys.path.insert(1, os.path.join(cmd_subfolder, "uns_sparkplugb"))
-from sparkplugb_enc_config import settings
+    sys.path.insert(0, cmd_subfolder)
+
+if uns_mqtt_folder not in sys.path:
+    sys.path.insert(2, uns_mqtt_folder)
+
+from uns_spb_mapper.sparkplugb_enc_config import settings
 
 is_configs_provided: bool = (
     os.path.exists(os.path.join(cmd_subfolder, "../conf/.secrets.yaml"))
@@ -77,12 +85,12 @@ def test_mqtt_config():
     assert (tls is None) or (os.path.isfile(tls.get(
         "ca_certs"))), f"Unable to find certificate at: {tls.get('ca_certs')}"
 
-    topics: str = settings.get("mqtt.topics", ["#"])
+    topics: str = settings.get("mqtt.topics", ["spBv1.0/#"])
     REGEX_TO_MATCH_TOPIC = r"^(\+|\#|.+/\+|[^#]+#|.*/\+/.*)$"
     for topic in topics:
         assert bool(
             re.fullmatch(REGEX_TO_MATCH_TOPIC, topic)
-        ), f"configuration 'mqtt.topic':{topic} is not a valid MQTT topic"
+        ), f"configuration 'mqtt.topics':{topics} has an valid MQTT topic topic:{topic}"
 
     keep_alive: float = settings.get("mqtt.keep_alive", 60)
     assert (keep_alive is None) or (
